chore(standardisation): move the action for report and submit utility grade to the correct controller
+ Update the link in thymleaf template
