+++ src/main/resources/templates/player/assignment/sequence/components/chat-gpt-evaluation/_chat-gpt-evaluation-report-modal.html
@@ -29,7 +29,7 @@
 
     <div class="content">
         <form th:id="${'report-form-' + evaluationId}"
-              th:action="@{/player/sequence/{sequenceId}/report-chat-gpt-evaluation(sequenceId=${chatGptEvaluationModel.sequenceId})}"
+              th:action="@{/chatGptEvaluation/sequence/{sequenceId}/report-chat-gpt-evaluation(sequenceId=${chatGptEvaluationModel.sequenceId})}"
               method="post">
             <div class="ui form">
                 <input type="hidden" name="evaluationId" th:value="${evaluationId}">
